refactored SuggestionDiffSummary
